Split up statestore module into package of modules
