+++ cognite/extractorutils/statestore/__init__.py
@@ -1,0 +1,12 @@
+from .classic import AbstractStateStore, LocalStateStore, NoStateStore, RawStateStore
+from .hashing import AbstractHashStateStore, LocalHashStateStore, RawHashStateStore
+
+__all__ = [
+    "AbstractStateStore",
+    "RawStateStore",
+    "LocalStateStore",
+    "NoStateStore",
+    "AbstractHashStateStore",
+    "RawHashStateStore",
+    "LocalHashStateStore",
+]
